Silence some Clang warnings
Specifically:

    match.c:142:31: warning: implicit conversion loses integer precision: 'VALUE' (aka 'unsigned long') to 'int' [-Wshorten-64-to-32]
        m.case_sensitive        = case_sensitive;
                                ~ ^~~~~~~~~~~~~~
    matcher.c:184:39: warning: implicit conversion loses integer precision: 'long' to 'int' [-Wshorten-64-to-32]
            thread_args[i].thread_count = thread_count;
                                        ~ ^~~~~~~~~~~~
    matcher.c:185:39: warning: implicit conversion loses integer precision: 'long' to 'int' [-Wshorten-64-to-32]
            thread_args[i].thread_index = i;
                                        ~ ^

+++ ruby/command-t/match.c
@@ -139,7 +139,7 @@ void calculate_match(VALUE str,
     m.max_score_per_char    = (1.0 / m.haystack_len + 1.0 / m.needle_len) / 2;
     m.always_show_dot_files = always_show_dot_files == Qtrue;
     m.never_show_dot_files  = never_show_dot_files == Qtrue;
-    m.case_sensitive        = case_sensitive;
+    m.case_sensitive        = (int)case_sensitive;
     m.recurse               = recurse == Qtrue;
 
     // calculate score
